+++ sso/verification/tests/test_models.py
@@ -4,6 +4,7 @@ import pytest
 from django.conf import settings
 from django.utils.timezone import now
 from freezegun import freeze_time
+import datetime
 
 from sso.verification.tests import factories
 
@@ -30,3 +31,9 @@ def test_verification_not_expired():
 def test_verification_model():
     verification_code = factories.VerificationCodeFactory()
     assert str(verification_code) == str(verification_code.user)
+
+
+@pytest.mark.django_db
+def test_verification_verified():
+    verification_code = factories.VerificationCodeFactory()
+    assert verification_code.date_verified == datetime.datetime(2021, 12, 29)